preload/dynamic_removing_preload.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    var t = async_test('Makes sure that dynamically removed preloaded resource stop downloading');
</script>
<body>
<script>
    var link = document.createElement("link");
    link.as = "script";
    link.rel = "preload";
    var url = "resources/dummy.js?pipe=trickle(d0.3)";
    link.href = url;
    document.body.appendChild(link);
    document.body.removeChild(link);
    window.addEventListener("load", t.step_func(function() {
        var entries = performance.getEntriesByName(new URL(url, location.href).href);
        assert_equals(entries.length, 0);
        t.done();
    }));
</script>
<script src="resources/dummy.js?pipe=trickle(d0.5)"></script>
</body>
